+++ apps/auth/messages/en.po
@@ -5,82 +5,6 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "X-Generator: next-intl\n"
 
-#: src/app/[locale]/(auth)/layout.tsx
-msgctxt "pRT4SY"
-msgid "Zoonk"
-msgstr "Zoonk"
-
-#: src/app/[locale]/(auth)/login/page.tsx
-msgctxt "4nOs3z"
-msgid "me@myorg.com"
-msgstr "me@myorg.com"
-
-#: src/app/[locale]/(auth)/login/page.tsx
-msgctxt "4wpUNc"
-msgid "Or"
-msgstr "Or"
-
-#: src/app/[locale]/(auth)/login/page.tsx
-#: src/app/[locale]/(auth)/otp/otp-form.tsx
-msgctxt "acrOoz"
-msgid "Continue"
-msgstr "Continue"
-
-#: src/app/[locale]/(auth)/login/page.tsx
-msgctxt "CoMQ1+"
-msgid "Continue with your email or a social account"
-msgstr "Continue with your email or a social account"
-
-#: src/app/[locale]/(auth)/login/page.tsx
-msgctxt "eN8MTJ"
-msgid "Sign in or create an account"
-msgstr "Sign in or create an account"
-
-#: src/app/[locale]/(auth)/login/page.tsx
-msgctxt "Ju9oB+"
-msgid "By clicking on Continue, you agree to our <terms>Terms of Service</terms> and <privacy>Privacy Policy</privacy>."
-msgstr "By clicking on Continue, you agree to our <terms>Terms of Service</terms> and <privacy>Privacy Policy</privacy>."
-
-#: src/app/[locale]/(auth)/login/page.tsx
-msgctxt "sy+pv5"
-msgid "Email"
-msgstr "Email"
-
-#: src/app/[locale]/(auth)/login/social-login.tsx
-msgctxt "BmecBA"
-msgid "Continue with Apple"
-msgstr "Continue with Apple"
-
-#: src/app/[locale]/(auth)/login/social-login.tsx
-msgctxt "FJ6r9E"
-msgid "Continue with Google"
-msgstr "Continue with Google"
-
-#: src/app/[locale]/(auth)/login/social-login.tsx
-msgctxt "pN2p0t"
-msgid "There was an error signing you in. Please try again or contact hello@zoonk.com"
-msgstr "There was an error signing you in. Please try again or contact hello@zoonk.com"
-
-#: src/app/[locale]/(auth)/otp/otp-form.tsx
-msgctxt "HkHvQv"
-msgid "The code you entered is incorrect. Please try again or contact hello@zoonk.com"
-msgstr "The code you entered is incorrect. Please try again or contact hello@zoonk.com"
-
-#: src/app/[locale]/(auth)/otp/page.tsx
-msgctxt "ChATeO"
-msgid "Change email"
-msgstr "Change email"
-
-#: src/app/[locale]/(auth)/otp/page.tsx
-msgctxt "LZHcmZ"
-msgid "Enter the code we sent to {email}:"
-msgstr "Enter the code we sent to {email}:"
-
-#: src/app/[locale]/(auth)/otp/page.tsx
-msgctxt "zx7CKB"
-msgid "Check your email"
-msgstr "Check your email"
-
 #: src/app/[locale]/callback/page.tsx
 msgctxt "7/cGw2"
 msgid "You're signed in!"
@@ -90,3 +14,79 @@ msgstr "You're signed in!"
 msgctxt "F8WbDs"
 msgid "You can close this window and return to your app."
 msgstr "You can close this window and return to your app."
+
+#: src/app/[locale]/layout.tsx
+msgctxt "dVItHD"
+msgid "Sign in to Zoonk"
+msgstr "Sign in to Zoonk"
+
+#: src/app/[locale]/login/page.tsx
+msgctxt "4nOs3z"
+msgid "me@myorg.com"
+msgstr "me@myorg.com"
+
+#: src/app/[locale]/login/page.tsx
+msgctxt "4wpUNc"
+msgid "Or"
+msgstr "Or"
+
+#: src/app/[locale]/login/page.tsx
+#: src/app/[locale]/otp/otp-form.tsx
+msgctxt "acrOoz"
+msgid "Continue"
+msgstr "Continue"
+
+#: src/app/[locale]/login/page.tsx
+msgctxt "CoMQ1+"
+msgid "Continue with your email or a social account"
+msgstr "Continue with your email or a social account"
+
+#: src/app/[locale]/login/page.tsx
+msgctxt "eN8MTJ"
+msgid "Sign in or create an account"
+msgstr "Sign in or create an account"
+
+#: src/app/[locale]/login/page.tsx
+msgctxt "Ju9oB+"
+msgid "By clicking on Continue, you agree to our <terms>Terms of Service</terms> and <privacy>Privacy Policy</privacy>."
+msgstr "By clicking on Continue, you agree to our <terms>Terms of Service</terms> and <privacy>Privacy Policy</privacy>."
+
+#: src/app/[locale]/login/page.tsx
+msgctxt "sy+pv5"
+msgid "Email"
+msgstr "Email"
+
+#: src/app/[locale]/login/social-login.tsx
+msgctxt "BmecBA"
+msgid "Continue with Apple"
+msgstr "Continue with Apple"
+
+#: src/app/[locale]/login/social-login.tsx
+msgctxt "FJ6r9E"
+msgid "Continue with Google"
+msgstr "Continue with Google"
+
+#: src/app/[locale]/login/social-login.tsx
+msgctxt "pN2p0t"
+msgid "There was an error signing you in. Please try again or contact hello@zoonk.com"
+msgstr "There was an error signing you in. Please try again or contact hello@zoonk.com"
+
+#: src/app/[locale]/otp/otp-form.tsx
+msgctxt "HkHvQv"
+msgid "The code you entered is incorrect. Please try again or contact hello@zoonk.com"
+msgstr "The code you entered is incorrect. Please try again or contact hello@zoonk.com"
+
+#: src/app/[locale]/otp/page.tsx
+msgctxt "ChATeO"
+msgid "Change email"
+msgstr "Change email"
+
+#: src/app/[locale]/otp/page.tsx
+msgctxt "LZHcmZ"
+msgid "Enter the code we sent to {email}:"
+msgstr "Enter the code we sent to {email}:"
+
+#: src/app/[locale]/otp/page.tsx
+msgctxt "zx7CKB"
+msgid "Check your email"
+msgstr "Check your email"
