+++ apps/auth/messages/es.po
@@ -5,88 +5,88 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "X-Generator: next-intl\n"
 
-#: src/app/[locale]/(auth)/layout.tsx
-msgctxt "pRT4SY"
-msgid "Zoonk"
-msgstr ""
-
-#: src/app/[locale]/(auth)/login/page.tsx
-msgctxt "4nOs3z"
-msgid "me@myorg.com"
-msgstr ""
-
-#: src/app/[locale]/(auth)/login/page.tsx
-msgctxt "4wpUNc"
-msgid "Or"
-msgstr ""
-
-#: src/app/[locale]/(auth)/login/page.tsx
-#: src/app/[locale]/(auth)/otp/otp-form.tsx
-msgctxt "acrOoz"
-msgid "Continue"
-msgstr ""
-
-#: src/app/[locale]/(auth)/login/page.tsx
-msgctxt "CoMQ1+"
-msgid "Continue with your email or a social account"
-msgstr ""
-
-#: src/app/[locale]/(auth)/login/page.tsx
-msgctxt "eN8MTJ"
-msgid "Sign in or create an account"
-msgstr ""
-
-#: src/app/[locale]/(auth)/login/page.tsx
-msgctxt "Ju9oB+"
-msgid "By clicking on Continue, you agree to our <terms>Terms of Service</terms> and <privacy>Privacy Policy</privacy>."
-msgstr ""
-
-#: src/app/[locale]/(auth)/login/page.tsx
-msgctxt "sy+pv5"
-msgid "Email"
-msgstr ""
-
-#: src/app/[locale]/(auth)/login/social-login.tsx
-msgctxt "BmecBA"
-msgid "Continue with Apple"
-msgstr ""
-
-#: src/app/[locale]/(auth)/login/social-login.tsx
-msgctxt "FJ6r9E"
-msgid "Continue with Google"
-msgstr ""
-
-#: src/app/[locale]/(auth)/login/social-login.tsx
-msgctxt "pN2p0t"
-msgid "There was an error signing you in. Please try again or contact hello@zoonk.com"
-msgstr ""
-
-#: src/app/[locale]/(auth)/otp/otp-form.tsx
-msgctxt "HkHvQv"
-msgid "The code you entered is incorrect. Please try again or contact hello@zoonk.com"
-msgstr ""
-
-#: src/app/[locale]/(auth)/otp/page.tsx
-msgctxt "ChATeO"
-msgid "Change email"
-msgstr ""
-
-#: src/app/[locale]/(auth)/otp/page.tsx
-msgctxt "LZHcmZ"
-msgid "Enter the code we sent to {email}:"
-msgstr ""
-
-#: src/app/[locale]/(auth)/otp/page.tsx
-msgctxt "zx7CKB"
-msgid "Check your email"
-msgstr ""
-
 #: src/app/[locale]/callback/page.tsx
 msgctxt "7/cGw2"
 msgid "You're signed in!"
-msgstr ""
+msgstr "¡Has iniciado sesión!"
 
 #: src/app/[locale]/callback/page.tsx
 msgctxt "F8WbDs"
 msgid "You can close this window and return to your app."
-msgstr ""
+msgstr "Puedes cerrar esta ventana y volver a tu aplicación."
+
+#: src/app/[locale]/layout.tsx
+msgctxt "dVItHD"
+msgid "Sign in to Zoonk"
+msgstr "Inicia sesión en Zoonk"
+
+#: src/app/[locale]/login/page.tsx
+msgctxt "4nOs3z"
+msgid "me@myorg.com"
+msgstr "yo@miempresa.com"
+
+#: src/app/[locale]/login/page.tsx
+msgctxt "4wpUNc"
+msgid "Or"
+msgstr "O"
+
+#: src/app/[locale]/login/page.tsx
+#: src/app/[locale]/otp/otp-form.tsx
+msgctxt "acrOoz"
+msgid "Continue"
+msgstr "Continuar"
+
+#: src/app/[locale]/login/page.tsx
+msgctxt "CoMQ1+"
+msgid "Continue with your email or a social account"
+msgstr "Continúa con tu correo o una cuenta social"
+
+#: src/app/[locale]/login/page.tsx
+msgctxt "eN8MTJ"
+msgid "Sign in or create an account"
+msgstr "Inicia sesión o crea una cuenta"
+
+#: src/app/[locale]/login/page.tsx
+msgctxt "Ju9oB+"
+msgid "By clicking on Continue, you agree to our <terms>Terms of Service</terms> and <privacy>Privacy Policy</privacy>."
+msgstr "Al hacer clic en Continuar, aceptas nuestros <terms>Términos de servicio</terms> y <privacy>Política de privacidad</privacy>."
+
+#: src/app/[locale]/login/page.tsx
+msgctxt "sy+pv5"
+msgid "Email"
+msgstr "Correo electrónico"
+
+#: src/app/[locale]/login/social-login.tsx
+msgctxt "BmecBA"
+msgid "Continue with Apple"
+msgstr "Continuar con Apple"
+
+#: src/app/[locale]/login/social-login.tsx
+msgctxt "FJ6r9E"
+msgid "Continue with Google"
+msgstr "Continuar con Google"
+
+#: src/app/[locale]/login/social-login.tsx
+msgctxt "pN2p0t"
+msgid "There was an error signing you in. Please try again or contact hello@zoonk.com"
+msgstr "Hubo un error al iniciar sesión. Intenta de nuevo o contacta a hello@zoonk.com"
+
+#: src/app/[locale]/otp/otp-form.tsx
+msgctxt "HkHvQv"
+msgid "The code you entered is incorrect. Please try again or contact hello@zoonk.com"
+msgstr "El código que ingresaste es incorrecto. Intenta de nuevo o contacta a hello@zoonk.com"
+
+#: src/app/[locale]/otp/page.tsx
+msgctxt "ChATeO"
+msgid "Change email"
+msgstr "Cambiar correo"
+
+#: src/app/[locale]/otp/page.tsx
+msgctxt "LZHcmZ"
+msgid "Enter the code we sent to {email}:"
+msgstr "Ingresa el código que enviamos a {email}:"
+
+#: src/app/[locale]/otp/page.tsx
+msgctxt "zx7CKB"
+msgid "Check your email"
+msgstr "Revisa tu correo"
