+++ apps/auth/messages/pt.po
@@ -5,88 +5,88 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "X-Generator: next-intl\n"
 
-#: src/app/[locale]/(auth)/layout.tsx
-msgctxt "pRT4SY"
-msgid "Zoonk"
-msgstr ""
-
-#: src/app/[locale]/(auth)/login/page.tsx
-msgctxt "4nOs3z"
-msgid "me@myorg.com"
-msgstr ""
-
-#: src/app/[locale]/(auth)/login/page.tsx
-msgctxt "4wpUNc"
-msgid "Or"
-msgstr ""
-
-#: src/app/[locale]/(auth)/login/page.tsx
-#: src/app/[locale]/(auth)/otp/otp-form.tsx
-msgctxt "acrOoz"
-msgid "Continue"
-msgstr ""
-
-#: src/app/[locale]/(auth)/login/page.tsx
-msgctxt "CoMQ1+"
-msgid "Continue with your email or a social account"
-msgstr ""
-
-#: src/app/[locale]/(auth)/login/page.tsx
-msgctxt "eN8MTJ"
-msgid "Sign in or create an account"
-msgstr ""
-
-#: src/app/[locale]/(auth)/login/page.tsx
-msgctxt "Ju9oB+"
-msgid "By clicking on Continue, you agree to our <terms>Terms of Service</terms> and <privacy>Privacy Policy</privacy>."
-msgstr ""
-
-#: src/app/[locale]/(auth)/login/page.tsx
-msgctxt "sy+pv5"
-msgid "Email"
-msgstr ""
-
-#: src/app/[locale]/(auth)/login/social-login.tsx
-msgctxt "BmecBA"
-msgid "Continue with Apple"
-msgstr ""
-
-#: src/app/[locale]/(auth)/login/social-login.tsx
-msgctxt "FJ6r9E"
-msgid "Continue with Google"
-msgstr ""
-
-#: src/app/[locale]/(auth)/login/social-login.tsx
-msgctxt "pN2p0t"
-msgid "There was an error signing you in. Please try again or contact hello@zoonk.com"
-msgstr ""
-
-#: src/app/[locale]/(auth)/otp/otp-form.tsx
-msgctxt "HkHvQv"
-msgid "The code you entered is incorrect. Please try again or contact hello@zoonk.com"
-msgstr ""
-
-#: src/app/[locale]/(auth)/otp/page.tsx
-msgctxt "ChATeO"
-msgid "Change email"
-msgstr ""
-
-#: src/app/[locale]/(auth)/otp/page.tsx
-msgctxt "LZHcmZ"
-msgid "Enter the code we sent to {email}:"
-msgstr ""
-
-#: src/app/[locale]/(auth)/otp/page.tsx
-msgctxt "zx7CKB"
-msgid "Check your email"
-msgstr ""
-
 #: src/app/[locale]/callback/page.tsx
 msgctxt "7/cGw2"
 msgid "You're signed in!"
-msgstr ""
+msgstr "Você está conectado!"
 
 #: src/app/[locale]/callback/page.tsx
 msgctxt "F8WbDs"
 msgid "You can close this window and return to your app."
-msgstr ""
+msgstr "Você pode fechar esta janela e voltar pro seu app."
+
+#: src/app/[locale]/layout.tsx
+msgctxt "dVItHD"
+msgid "Sign in to Zoonk"
+msgstr "Entrar no Zoonk"
+
+#: src/app/[locale]/login/page.tsx
+msgctxt "4nOs3z"
+msgid "me@myorg.com"
+msgstr "eu@minhaempresa.com"
+
+#: src/app/[locale]/login/page.tsx
+msgctxt "4wpUNc"
+msgid "Or"
+msgstr "Ou"
+
+#: src/app/[locale]/login/page.tsx
+#: src/app/[locale]/otp/otp-form.tsx
+msgctxt "acrOoz"
+msgid "Continue"
+msgstr "Continuar"
+
+#: src/app/[locale]/login/page.tsx
+msgctxt "CoMQ1+"
+msgid "Continue with your email or a social account"
+msgstr "Continue com seu email ou uma conta social"
+
+#: src/app/[locale]/login/page.tsx
+msgctxt "eN8MTJ"
+msgid "Sign in or create an account"
+msgstr "Entre ou crie uma conta"
+
+#: src/app/[locale]/login/page.tsx
+msgctxt "Ju9oB+"
+msgid "By clicking on Continue, you agree to our <terms>Terms of Service</terms> and <privacy>Privacy Policy</privacy>."
+msgstr "Ao clicar em Continuar, você concorda com nossos <terms>Termos de Serviço</terms> e <privacy>Política de Privacidade</privacy>."
+
+#: src/app/[locale]/login/page.tsx
+msgctxt "sy+pv5"
+msgid "Email"
+msgstr "Email"
+
+#: src/app/[locale]/login/social-login.tsx
+msgctxt "BmecBA"
+msgid "Continue with Apple"
+msgstr "Continuar com Apple"
+
+#: src/app/[locale]/login/social-login.tsx
+msgctxt "FJ6r9E"
+msgid "Continue with Google"
+msgstr "Continuar com Google"
+
+#: src/app/[locale]/login/social-login.tsx
+msgctxt "pN2p0t"
+msgid "There was an error signing you in. Please try again or contact hello@zoonk.com"
+msgstr "Houve um erro ao conectar você. Por favor, tente novamente ou entre em contato com hello@zoonk.com"
+
+#: src/app/[locale]/otp/otp-form.tsx
+msgctxt "HkHvQv"
+msgid "The code you entered is incorrect. Please try again or contact hello@zoonk.com"
+msgstr "O código que você digitou está incorreto. Por favor, tente novamente ou entre em contato com hello@zoonk.com"
+
+#: src/app/[locale]/otp/page.tsx
+msgctxt "ChATeO"
+msgid "Change email"
+msgstr "Alterar email"
+
+#: src/app/[locale]/otp/page.tsx
+msgctxt "LZHcmZ"
+msgid "Enter the code we sent to {email}:"
+msgstr "Digite o código que enviamos para {email}:"
+
+#: src/app/[locale]/otp/page.tsx
+msgctxt "zx7CKB"
+msgid "Check your email"
+msgstr "Confira seu email"
