+++ apps/auth/src/app/[locale]/otp/loading.tsx
