+++ i18n.json
@@ -3,6 +3,7 @@
   "buckets": {
     "po": {
       "include": [
+        "apps/auth/messages/[locale].po",
         "apps/editor/messages/[locale].po",
         "apps/main/messages/[locale].po"
       ]
