+++ i18n.lock
@@ -166,3 +166,21 @@ checksums:
     Learn/singular: 6559e2075ab6947a1d4fa43e19a5d66b
     Checking%20if%20you're%20logged%20in.../singular: b50f05bfcb51cad6913150273e551291
     You%20need%20to%20be%20logged%20in%20to%20access%20this%20page./singular: 759ca522663d8f628718332f756dd6c4
+  61a308caf583f5835f54302028bdae73:
+    You're%20signed%20in!/singular: 7f23ad13626e60f22e13b8777458750d
+    You%20can%20close%20this%20window%20and%20return%20to%20your%20app./singular: 06f96197fc40b30e288b1a9603d15825
+    Sign%20in%20to%20Zoonk/singular: e382c90df1855628b2ea78c3d8a62ea9
+    me%40myorg.com/singular: 261d9ee525c108801f822ecb3fa7b8e8
+    Or/singular: 98639bcbc7b4889bae62ad0ebde740e1
+    Continue/singular: 3cfba90b4600131e82fc4260c568d044
+    Continue%20with%20your%20email%20or%20a%20social%20account/singular: 1f2b79de8d5cb3a40df757ab1b50db55
+    Sign%20in%20or%20create%20an%20account/singular: b29f31bce3d279137572901bf10d6e5f
+    By%20clicking%20on%20Continue%2C%20you%20agree%20to%20our%20%3Cterms%3ETerms%20of%20Service%3C%2Fterms%3E%20and%20%3Cprivacy%3EPrivacy%20Policy%3C%2Fprivacy%3E./singular: 5056e412e67c45780a80009328ad062d
+    Email/singular: e7f34943a0c2fb849db1839ff6ef5cb5
+    Continue%20with%20Apple/singular: 8b13538d47736d8b7b0b0edbbeb9c741
+    Continue%20with%20Google/singular: 296841ac969a7f5d0fffc144f2488bbb
+    There%20was%20an%20error%20signing%20you%20in.%20Please%20try%20again%20or%20contact%20hello%40zoonk.com/singular: abb16bd44edbd2ace037414c5ab9567b
+    The%20code%20you%20entered%20is%20incorrect.%20Please%20try%20again%20or%20contact%20hello%40zoonk.com/singular: b437ceb9d111fd91320f586ee73f7024
+    Change%20email/singular: de3d14d915838e1eb5b04e620de35114
+    Enter%20the%20code%20we%20sent%20to%20%7Bemail%7D%3A/singular: 939208aaf736312126ef0fe08cb1faa3
+    Check%20your%20email/singular: c51aa8640fe15af146f3e2dd4602e0c1
